pyproject.toml
[tool.poetry]
name = "insighter"
version = "0.1.0"
description = ""
authors = ["artemprod <artemprod1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.4.1"
docker = "^7.0.0"
environs = "^11.0.0"
ffmpeg-python = "^0.2.0"
langsmith = "^0.1.27"
mongoengine = "^0.28.2"
smart-open = "^7.0.1"
tiktoken = "^0.6.0"
nltk = "^3.8.1"
filetype = "^1.2.0"
paramiko = "^3.4.0"
pydub = "^0.25.1"
aiofiles = "^23.2.1"
redis = "^5.0.3"
ruff = "^0.3.3"
pre-commit = "^3.6.2"
loguru = "^0.7.2"
notifiers = "^1.3.3"
logtail-python = "^0.2.10"
chardet = "^5.2.0"
pypdf2 = "^3.0.1"
mutagen = "^1.47.0"
retry = "^0.9.2"
openai = "^1.14.2"
langchain-openai = "^0.1.0"
langchain-core = "^0.1.33"
langchain-community = "^0.0.29"
langchain = "^0.1.13"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 120

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint]
# See definitions https://docs.astral.sh/ruff/rules/
select = [
    "ASYNC", # flake8-async
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "E",     # pycodestyle errors
    "F",     # Pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "SIM",   # flake8-simplify
    "T20",   # flake8-insighter_logger.info
    "TRIO",  # flake8-trio
    "S",     # flake8-bandit
#    "EM",    # flake8-errmsg
    "LOG",  # flake8-logging
#    "RUF",  # Ruff-specific rules
]
fixable = ["ALL"]

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

